Uploaded Lab03
Lab description is to create a BookApp that has a Bottom Tab Navigation (BookList and BookBorrowed) BookList shows a list of books it's title, author, and availability. The user can click each item on the list and will be redirected to Book Details page which shows further details of the book and a function button that allows the user to borrow the book if it hasn't reached the maximum book borrowing count (3). The BookBorrowed screen shows books borrowed and has a button functionality to return the book.